Create segment_tree_standard_template
problem - 
Description
You've given an array A of N integers. Your task is to support the following queries.

0 x y : Modify the number at index x to y.
1 l r : Print the count the number of even numbers in range l to r inclusive.
2 l r : Print the count the number of odd numbers in range l to r inclusive.
Input Format
The first input line has an integer N, the number of elements.
The second line has N integers A1, A2, …, AN: the array values.
The third line has an integer Q, the number of queries.
Each of the next Q lines contains the query of one of the three types described in the statement. 

Output Format
Print the output of query of type 1 and 2 on a new line.

Constraints
1 ≤ N, Q ≤ 105
1 ≤ l ≤ r ≤ N
0 ≤ Ai ≤ 109
1 ≤ x ≤ N
0 ≤ y ≤ 109

Sample Input 1
6
1 2 3 4 5 6
4
1 2 5
2 1 4
0 5 4
1 1 6
Sample Output 1
2
2
4